requirements.txt
blinker==1.5
click==8.0.1
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.0.1
MarkupSafe==2.0.1
nulltype==2.3.1
plaid-python==11.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.13
Werkzeug==2.0.1
